The client now tries two hardcoded servers for HTTP downloads first.
If it can't find the file at http://red.planetarena.org/sv_downloadurl or
http://martianbackup.com/sv_downloadurl, *then* it will try the game server's
sv_downloadurl, and *then* if that fails, an old-fashioned UDP download.

If the server's sv_downloadurl is the same as one of the hardcoded URLs, it
will detect this and still only try that URL once.


git-svn-id: svn://svn.icculus.org/alienarena/trunk@3847 05fc5d3f-ce2a-0410-9b80-abac23765f80