Introduce and begin using CM_FastTrace funcion.
This function is like CM_BoxTrace, but simply returns true if the path is
clear, without finding the precise locations of any obstructions. If you're
calling CM_BoxTrace just to check if the trace fraction is 1.0, you should
switch to CM_FastTrace.

It appears to be significantly faster than CM_BoxTrace when tracing through
BSPs, but it's not currently faster for terrain maps. I have some ideas for
that though.



git-svn-id: svn://svn.icculus.org/alienarena/trunk@4449 05fc5d3f-ce2a-0410-9b80-abac23765f80