major overhaul of shadowmaps.
All sun shadows (vegetation and entity) go in the same depth buffer. This is
currently the only type of shadow that can be cast on terrain-- meaning we
get vegetation shadows on terrain now.

Fix wrapping-around and repeating of shadows on terrain.

There are now separate depth buffers for sun and non-sun static shadows. Some
surface types (meshes) support both at once-- meaning we get vegetation
shadows on meshes now.

Move the stencil FBO init to r_postprocess.c since that's the only file that
uses that FBO anymore.

Exclude sun lights from light groups and (temporarily) disable darkening of
meshes when sun is visible, so as to segregate sun and non-sun static shadows.

It's now possible for an entity to cast two static shadows, sun and non-sun.
The non-sun one is faded out significantly if the sun one is visible.


git-svn-id: svn://svn.icculus.org/alienarena/trunk@4688 05fc5d3f-ce2a-0410-9b80-abac23765f80