Add initial, very basic visualizer program for libgarland mesh simplification.
This program allows you to watch the terrain mesh simplification as it
happens, pause and resume it, and skip forward a certain number of steps.

Currently only tested on Linux, requires SDL 1.x, has its own Makefile.

Required some changes to the main game codebase so the code could be reused in
this program.


git-svn-id: svn://svn.icculus.org/alienarena/trunk@4744 05fc5d3f-ce2a-0410-9b80-abac23765f80