remove unused acebot_compress.c.
None of the code in this file was actually used. It wasn't mentioned in the
Unix makefiles but removing this file may temporarily break the Visual Studio
build.

That file was scary from a copyright standpoint anyway, since the original
acebot author just lifted that code from somewhere without remembering who to
credit.


git-svn-id: svn://svn.icculus.org/alienarena/trunk@4773 05fc5d3f-ce2a-0410-9b80-abac23765f80