game: remove antilag cvar, default antilagged projectiles on.
Antilag for hitscan weapons is so well-tested at this point, there's really no
reason ever to turn it off. Giving server hosts the option to turn it off will
only cause an inconsistent experience.

Meanwhile, we want to encourage server admins to run antilagged projectiles,
which will allow more testing of the code.


git-svn-id: svn://svn.icculus.org/alienarena/trunk@4874 05fc5d3f-ce2a-0410-9b80-abac23765f80