game: p_client: bulk of spectator/team select overhaul done.
This is using the design described here:
    http://red.planetarena.org/forum/showthread.php?tid=1920
This includes adding the "teamselect" console command, which was added server-
side so even old clients will be able to use it. Spectator and team selection
functionality is all complete and, in my testing, it seems to work the same as
before (minus some bugs that got fixed.) This includes even obscure stuff like
spectator passwords. Unfortunately, I've had to temporarily disable duel mode,
but I'll re-enable it soon.

This commit should also fix the "immortality hack" that got reported a while
back: http://red.planetarena.org/forum/showthread.php?tid=1856

From now on, throughout the code, instead of checking whether a player is
"spectating," you should check whether a player is "participating." I wrote a
new player_participating function for convenience, or you could check
ent->client->resp.participation directly.

A non-participating player *might* be spectating... or he might be picking a
team in CTF. Or, it could be something else altogether. I plan on handling duel
waiting through this system. In my opinion, this system would also be a great
way to handle class selection in Tactical.

Please do not *modify* ent->client->resp.participation directly. Instead, you
should arrange it so that when a player's participation status needs to change,
the player_desired_participation function (p_client.c) will start to return a
different value. As soon as that function starts returning a different value,
the rest of the code will do all the work of changing everything for you.

High risk of regressions for this commit.


git-svn-id: svn://svn.icculus.org/alienarena/trunk@5076 05fc5d3f-ce2a-0410-9b80-abac23765f80