game: cleanup reward powerup system.
Once a player crosses the reward point threshold, all the buyable powerups are
no longer put into his inventory.

Items can now have an IT_BUYABLE flag, and the item->use callback is no longer
expected to handle reward points at all.

Instead, when someone uses the "use" command on a buyable item which is not
in the inventory, the code for the command itself checks if the player is
beyond the reward points threshold and if so allows the player to use it
anyway.

Now it's possible to make more powerups buyable just by including the
IT_BUYABLE flag and nothing else (although there's some extra work if you
want the AI to buy it.)


git-svn-id: svn://svn.icculus.org/alienarena/trunk@5089 05fc5d3f-ce2a-0410-9b80-abac23765f80