unix: don't grab the mouse during X11 setup.
This prevents the game from grabbing focus during startup and GL init, which
can render the system unusable during that period or evem render it unusable
altogether if GL init hangs.

The default starting state of the game (main menu) shouldn't have the mouse
grabbed anyway, so grabbing it for a split second before the menu comes up
is just an annoyance.

The existing IN_Frame function need not be modified, because it will already
grab the mouse if appropriate every frame. So as soon as frames start getting
rendered, it will grab the mouse if appropriate.


git-svn-id: svn://svn.icculus.org/alienarena/trunk@5548 05fc5d3f-ce2a-0410-9b80-abac23765f80